+++ src/main/java/mate/academy/springboot/datajpa/service/ProductService.java
@@ -1,10 +1,10 @@
 package mate.academy.springboot.datajpa.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import mate.academy.springboot.datajpa.model.Product;
 
 public interface ProductService {
-
     Product add(Product category);
 
     Product findById(Long productId);
@@ -13,7 +13,7 @@ public interface ProductService {
 
     Product update(Product product);
 
-    List<Product> findByPriceBetween(Float priceFrom, Float priceTo);
+    List<Product> findByPriceBetween(BigDecimal priceFrom, BigDecimal priceTo);
 
     List<Product> findByCategories(List<String> categories);
 }
